Nucleus/requirements.txt
fastapi[all]
uvicorn[standard]
python-decouple
pymssql
requests
websockets
asyncio